Make MTU a uint16
A MTU can never be negative so this removes a class of runtime errors.

+++ codecs/opus_packet.go
@@ -4,7 +4,7 @@ package codecs
 type OpusPayloader struct{}
 
 // Payload fragments an Opus packet across one or more byte arrays
-func (p *OpusPayloader) Payload(mtu int, payload []byte) [][]byte {
+func (p *OpusPayloader) Payload(mtu uint16, payload []byte) [][]byte {
 	if payload == nil {
 		return [][]byte{}
 	}
